# Speed-up span Id generation