Use the RandomNumberGenerator.Next instead of Guid.New to generate span Ids. Besides saving the allocation of the array Guid.New is pretty slow on Linux.

Linux:
|     Method |        Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------- |------------:|----------:|----------:|-------:|------:|------:|----------:|
|     NextId |    12.52 ns |  0.317 ns |  0.900 ns |      - |     - |     - |         - |
| GenerateId | 1,611.62 ns | 31.538 ns | 30.975 ns | 0.0095 |     - |     - |      40 B |

Windows:
|     Method |      Mean |     Error |    StdDev |    Median |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------- |----------:|----------:|----------:|----------:|-------:|------:|------:|----------:|
|     NextId |  9.495 ns | 0.2111 ns | 0.4498 ns |  9.481 ns |      - |     - |     - |         - |
| GenerateId | 80.733 ns | 1.6216 ns | 4.2718 ns | 79.465 ns | 0.0095 |     - |     - |      40 B |